// frontend/src/api/generated/models/transferResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * FastTrans API
 * OpenAPI spec version: 1.0.0
 */

export interface TransferResponse {
  id?: string;
  fromAccountRef?: string;
  toAccountRef?: string;
  amount?: number;
  currency?: string;
  status?: string;
  reason?: string;
  createdAt?: string;
}